# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TodoTalk messenger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List chat rooms I'm in
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Room list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatRoom"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a chat room
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatRoom"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room details
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatRoom"
  /rooms/{id}/leave:
    post:
      operationId: leaveRoom
      tags: [rooms]
      summary: Leave a room
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Left
  /rooms/{id}/mute:
    patch:
      operationId: toggleMuteRoom
      tags: [rooms]
      summary: Toggle mute for a room
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MuteInput"
      responses:
        "200":
          description: Mute toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatRoomMember"
  /rooms/{id}/members:
    get:
      operationId: listRoomMembers
      tags: [rooms]
      summary: List room members
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicUser"
    post:
      operationId: inviteRoomMembers
      tags: [rooms]
      summary: Invite additional friends to a group room
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMembersInput"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatRoom"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PublicUser:
      type: object
      required: [id, nickname, accountKind]
      properties:
        id:
          type: string
        nickname:
          type: string
        accountKind:
          type: string
          enum: [user, official, system]
        friendAlias:
          type: ["string", "null"]
          description: Current viewer's saved name for this user, when available.
        displayName:
          type: string
          description: friendAlias when set, otherwise nickname.
        profileImageUrl:
          type: ["string", "null"]
        statusMessage:
          type: ["string", "null"]
...
    ChatRoom:
      type: object
      required: [id, type, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
        name:
          type: ["string", "null"]
        ownerId:
          type: ["string", "null"]
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        lastMessageSeq:
          type: integer
        pinnedMessageId:
          type: ["string", "null"]
        pinnedMessage:
          anyOf:
            - $ref: "#/components/schemas/MessageSummary"
            - type: "null"
        unreadCount:
          type: integer
        firstUnreadMessageId:
          type: ["string", "null"]
        muted:
          type: boolean
        createdAt:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/PublicUser"
    MessageSummary:
      type: object
      required: [id, senderId, type, content, createdAt]
      properties:
        id:
          type: string
        senderId:
          type: string
        senderName:
          type: ["string", "null"]
        type:
          type: string
        content:
          type: string
        createdAt:
          type: string
        deletedAt:
          type: ["string", "null"]
    ChatRoomMember:
      type: object
      required: [id, roomId, userId, joinedAt, muted]
      properties:
        id:
          type: string
        roomId:
          type: string
        userId:
          type: string
        joinedAt:
          type: string
        muted:
          type: boolean
    RoomInput:
      type: object
      required: [type, memberIds]
      properties:
        type:
          type: string
          enum: [direct, group]
        name:
          type: ["string", "null"]
          maxLength: 120
        memberIds:
          type: array
          maxItems: 100
          items:
            type: string
            format: uuid
    InviteMembersInput:
      type: object
      required: [memberIds]
      properties:
        memberIds:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: string
            format: uuid
...
    MuteInput:
      type: object
      required: [muted]
      properties:
        muted:
          type: boolean
...